requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
fastjsonschema==2.16.3
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lazy-loader==0.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mistune==2.0.5
multidict==6.0.4
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.0
pycryptodome==3.17
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
pyzmq==25.0.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
soupsieve==2.4
tabulate==0.9.0
tifffile==2023.2.28
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
webencodings==0.5.1
yarl==1.8.2
zipp==3.15.0
